refactor(frontend): extract Pipeline card and modals into pipeline/
Pipeline.jsx (740 lines) keeps data loading, filters, and action
dispatch; JobCard and the four modals (Add Job, Edit Resume, Edit Cover
Letter, Templates) move to frontend/src/screens/pipeline/ with shared
modal primitives. AddJobModal now reuses the same Modal wrapper it
duplicated inline. No behavior changes.
